10/24 jest 대시보드 폴더 변경 및 적용
